refactor(project-structure): Refactor the directory to one source folder and one webpack.
