--- src-v2/manifest.json
@@ -1,64 +1,0 @@
-{
-  "manifest_version": 3,
-  "name": "FreeScribe",
-  "version": "1.0",
-  "description": "Record and play audio from tab and microphone",
-  "permissions": [
-    "activeTab",
-    "contextMenus",
-    "offscreen",
-    "storage",
-    "scripting",
-    "tabs",
-    "notifications"
-  ],
-  "icons": {
-    "16": "icons/icon16.png",
-    "32": "icons/icon32.png",
-    "48": "icons/icon48.png",
-    "128": "icons/icon128.png"
-  },
-  "host_permissions": [
-    "<all_urls>"
-  ],
-  "options_page": "options.html",
-  "background": {
-    "service_worker": "background.js"
-  },
-  "action": {
-    "default_title": "FreeScribe"
-  },
-  "content_security_policy": {
-    "extension_pages": "script-src 'self' 'wasm-unsafe-eval'"
-  },
-  "commands": {
-    "_execute_action": {
-      "suggested_key": {
-        "default": "Alt+Shift+F"
-      },
-      "description": "Open FreeScribe Extension"
-    },
-    "start_stop_recording": {
-      "suggested_key": "Alt+Shift+R",
-      "description": "Start or Stop Recording"
-    },
-    "configure": {
-      "suggested_key": "Alt+Shift+Comma",
-      "description": "Open FreeScribe Configuration"
-    }
-  },
-  "web_accessible_resources": [
-    {
-      "resources": [
-        "*.html",
-        "*.css",
-        "*.svg",
-        "*.png"
-      ],
-      "matches": [
-        "http://*/*",
-        "https://*/*"
-      ]
-    }
-  ]
-}
