--- src-v2/popup.html
@@ -1,176 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8"/>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
-    <title>FreeScribe</title>
-    <link rel="stylesheet" href="main2.css"/>
-    <link rel="stylesheet" href="fontawesome.min.css"/>
-    <style>
-.loading-overlay {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: rgba(255, 255, 255, 0.9);
-    display: none;
-    justify-content: center;
-    align-items: center;
-    z-index: 1000;
-    flex-direction: column;
-}
-
-.loading-content {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 1rem;
-}
-
-.loading-text {
-    font-size: 1.2rem;
-    color: #333;
-}
-</style>
-    <script src="jquery-3.7.1.min.js"></script>
-    <script src="toastr.min.js"></script>
-    <script src="bootstrap.bundle.min.js"></script>
-</head>
-<body>
-<div id="loadingOverlay" class="loading-overlay">
-    <div class="loading-content">
-        <div class="spinner-border text-primary" style="width: 3rem; height: 3rem;" role="status">
-            <span class="visually-hidden">Loading...</span>
-        </div>
-        <p id="loadingText" class="loading-text">Loading models...</p>
-    </div>
-</div>
-<div class="popup-container">
-    <div class="container fade-in">
-        <h3>Select Audio Input</h3>
-        <select id="audioInputSelect" class="mb-3"></select>
-
-        <div id="volumeBar" class="mb-3">
-            <div id="volumeLevel"></div>
-        </div>
-
-        <div class="d-flex justify-content-between mb-3">
-            <button id="recordButton" class="flex-grow-1 me-2 btn-success">
-                <i class="fas fa-microphone"></i> Record
-            </button>
-
-            <button
-                    id="stopButton"
-                    style="display: none"
-                    class="flex-grow-1 me-2 btn-danger"
-            >
-                <i class="fas fa-stop"></i> Stop
-            </button>
-
-            <button
-                    id="pauseButton"
-                    disabled="true"
-                    class="flex-grow-1 me-2 btn-primary"
-            >
-                <i class="fas fa-pause"></i> Pause
-            </button>
-
-            <button
-                    id="resumeButton"
-                    style="display: none"
-                    class="flex-grow-1 me-2 btn-success"
-            >
-                <i class="fas fa-play"></i> Resume
-            </button>
-        </div>
-
-        <div id="errorMessage" class="alert alert-danger" role="alert" style="display: none;"></div>
-
-        <div class="textarea-container mb-3">
-          <textarea
-                  id="userInput"
-                  rows="5"
-                  class="form-control"
-                  placeholder=""
-          ></textarea>
-            <div id="s2t-loader" class="text-center loader">
-                <div class="spinner-border" role="status">
-                    <span class="sr-only">Loading...</span>
-                </div>
-            </div>
-        </div>
-
-        <button id="generateNotesButton" class="mb-3">
-            <i class="fas fa-file-alt"></i> Generate Note
-        </button>
-
-        <pre id="notes" class="mt-3"></pre>
-
-        <!-- Copy to Clipboard Button -->
-        <button id="copyNotesButton" class="btn btn-secondary mb-3">
-            <i class="fas fa-copy"></i> Copy Notes
-        </button>
-
-        <div class="text-center mt-4">
-            <a href="#" id="toggleConfig" class="text-muted settings">
-                <i class="fas fa-cog"></i> Configuration
-            </a>
-        </div>
-
-        <div class="text-center mt-4">
-            <a href="#" id="showHistory" class="text-muted settings">
-                <i class="fas fa-history"></i> History
-            </a>
-        </div>
-    </div>
-</div>
-
-<div id="loadingStatus" class="hidden">
-    <p id="loadingMessage">Loading...</p>
-    <div id="progressContainer"></div>
-</div>
-
-<div class="text-center mt-4">
-    <p>
-        For more information, click
-        <a href="welcome.html" target="_blank">here</a>.
-    </p>
-</div>
-<div
-        class="modal fade"
-        id="historyModel"
-        tabindex="-1"
-        aria-labelledby="historyTitle"
-        aria-hidden="true"
->
-    <div class="modal-dialog modal-dialog-scrollable">
-        <div class="modal-content">
-            <div class="modal-header">
-                <h1 class="modal-title fs-5" id="historyTitle">History</h1>
-                <button
-                        type="button"
-                        class="btn-close"
-                        data-bs-dismiss="modal"
-                        aria-label="Close"
-                ></button>
-            </div>
-            <div class="modal-body">
-                <div class="accordion" id="historyAccordion"></div>
-            </div>
-            <div class="modal-footer">
-                <button
-                        type="button"
-                        class="btn btn-secondary"
-                        data-bs-dismiss="modal"
-                >
-                    Close
-                </button>
-            </div>
-        </div>
-    </div>
-</div>
-
-<script type="module" src="popup.js"></script>
-</body>
-</html>
