--- src-v2/welcome.html
@@ -1,166 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Welcome to FreeScribe</title>
-    <link rel="stylesheet" href="bootstrap.css" />
-    <link rel="stylesheet" href="main.css" />
-  </head>
-  <body class="free-scribe">
-    <div class="welcome-container">
-      <div class="content">
-        <h1 class="mb-4">Welcome to FreeScribe</h1>
-        <p class="lead mb-4">
-          FreeScribe is a powerful browser extension that helps you transcribe
-          audio, generate notes, and manage your medical documentation
-          effortlessly.
-        </p>
-        <div class="pin-extension-note">
-          📌 For the best experience, please pin the FreeScribe extension to your browser toolbar.
-        </div>
-        <h2 class="mb-3">Key Features:</h2>
-        <ul class="feature-list mb-4">
-          <li>
-            <strong>Flexible Processing Options</strong>
-            <p>
-              Choose between <strong>local</strong> or <strong>server-based</strong> transcription and LLM
-              processing to suit your needs
-            </p>
-            <aside class="webgpu-note" role="note">
-              ⚙️ When using <strong>local</strong> transcription and LLM, please check and enable <strong>Unsafe WebGPU Support</strong> from <a href="chrome://flags/#enable-unsafe-webgpu">chrome://flags/#enable-unsafe-webgpu</a>
-              and <strong>use graphics acceleration when available</strong> from <a href="chrome://settings/system">chrome://settings/system</a>.
-              <strong>Please copy these links and paste them into your browser's address bar.</strong>
-            </aside>
-          </li>
-          <li>
-            <strong>Real-time Audio Transcription</strong>
-            <p>
-              Convert speech to text instantly with high accuracy and automatic
-              punctuation
-            </p>
-          </li>
-          <li>
-            <strong>AI-powered Note Generation</strong>
-            <p>
-              Transform raw transcriptions into structured medical notes using
-              advanced AI processing
-            </p>
-          </li>
-          <li>
-            <strong>Customizable Templates</strong>
-            <p>
-              Use and modify pre-built templates or create your own custom
-              prompts for different medical scenarios
-            </p>
-          </li>
-          <li>
-            <strong>Pre/Post Processing</strong>
-            <p>
-              Apply custom rules and formatting to your notes before and after
-              AI processing
-            </p>
-          </li>
-          <li>
-            <strong>History Management</strong>
-            <p>
-              Track, search, and manage your transcription and note history with
-              easy access to past records
-            </p>
-          </li>
-          <li>
-            <strong>Device Flexibility</strong>
-            <p>
-              Support for multiple audio input devices and microphone
-              configurations
-            </p>
-          </li>
-          <li>
-            <strong>Security First</strong>
-            <p>
-              Process all data locally with optional cloud features, ensuring
-              complete control over sensitive information
-            </p>
-          </li>
-          <li>
-            <strong>Easy Export</strong>
-            <p>Copy your notes in just one click</p>
-          </li>
-        </ul>
-
-        <h2 class="mb-3">Keyboard Shortcuts:</h2>
-        <ul class="shortcut-list mb-4">
-          <li>
-            Open FreeScribe Extension:
-            <code>Alt + Shift + F</code>
-          </li>
-          <li>Start/Stop Recording: <code>Alt + Shift + R</code></li>
-          <li>Configure: <code>Alt + Shift + Comma (,)</code></li>
-        </ul>
-
-        <p class="mb-4">
-          To get started, click on the FreeScribe icon in your browser toolbar.
-          You can record audio, transcribe it, generate notes, and manage your
-          history all from one convenient interface.
-        </p>
-
-        <p class="mb-4">
-          For the best experience, we recommend configuring FreeScribe to work
-          with your preferred settings <a href="#" id="configureLink" class="btn-lg">here</a>.
-        </p>
-
-        <h2 class="mb-3">Installations (Optional):</h2>
-        <p class="mb-4">
-          If using <strong>server-based</strong> processing, you will need to install the following dependencies:
-        </p>
-        <ul class="mb-4">
-          <li>
-            <a
-              href="https://github.com/ClinicianFOCUS/local-llm-container"
-              target="_blank"
-              >Local LLM Container</a
-            >: Required for running the local language model. Follow the
-            installation instructions in the repository.
-          </li>
-          <li>
-            <a
-              href="https://github.com/ClinicianFOCUS/speech2text-container"
-              target="_blank"
-              >Speech to Text Container</a
-            >: Needed for audio transcription. Please refer to the repository
-            for installation details.
-          </li>
-          <li>
-            <a
-                    href="https://github.com/ClinicianFOCUS/FreeScribe"
-                    target="_blank"
-            >FreeScribe</a
-            >: The main project repository. Please follow the installation
-            instructions provided in the repository.
-          </li>
-        </ul>
-
-        <p class="mb-4">
-          For detailed installation steps, please refer to the README files in
-          each of the above repositories.
-        </p>
-
-        <div class="mt-4 text-center">
-          <p>
-            License: <strong>GNU Affero General Public License v3.0</strong>
-          </p>
-          <p>
-            For detailed documentation, visit our
-            <a
-              href="https://github.com/ClinicianFOCUS/freescribe-copilot"
-              target="_blank"
-              >GitHub repository</a
-            >
-            to learn more about the project or report issues.
-          </p>
-        </div>
-      </div>
-    </div>
-    <script src="welcome.js"></script>
-  </body>
-</html>
