+++ src/options.js
@@ -1,8 +1,9 @@
-import { loadConfig, saveConfig } from "./config.js";
-import { isValidUrl } from "./helpers.js";
+import {loadConfig, saveConfig} from "./config.js";
+import {isValidUrl} from "./helpers.js";
 
 // variables for the configuration settings
 let config;
+const customModelKey = "custom";
 
 // Function: getOptionsFromArray - Get the html options from an array of models
 // Map the array of models to an array of option elements and join them
@@ -59,21 +60,32 @@ function loadRemoteModelsOnSettingsChange() {
 function showConfig() {
     // Transcription settings
     document.getElementById("transcriptionLocal").checked = config.TRANSCRIPTION_LOCAL;
-    toggleTranscriptionSettings();
 
     const transcriptionLocalModelSelect = document.getElementById("transcriptionLocalModel");
-    transcriptionLocalModelSelect.innerHTML = getOptionsFromArray(config.TRANSCRIPTION_LOCAL_MODELS);
-    transcriptionLocalModelSelect.value = config.TRANSCRIPTION_LOCAL_MODEL;
+    transcriptionLocalModelSelect.innerHTML = getOptionsFromArray([...config.TRANSCRIPTION_LOCAL_MODELS, customModelKey]);
+
+    if (config.TRANSCRIPTION_LOCAL_MODELS.includes(config.TRANSCRIPTION_LOCAL_MODEL)) {
+        transcriptionLocalModelSelect.value = config.TRANSCRIPTION_LOCAL_MODEL;
+    } else {
+        transcriptionLocalModelSelect.value = customModelKey;
+    }
 
     document.getElementById("transcriptionUrl").value = config.TRANSCRIPTION_URL;
     document.getElementById("transcriptionApiKey").value = config.TRANSCRIPTION_API_KEY;
+
+    toggleTranscriptionSettings();
 
     // LLM settings
     document.getElementById("llmLocal").checked = config.LLM_LOCAL;
 
     const llmLocalModelSelect = document.getElementById("llmLocalModel");
-    llmLocalModelSelect.innerHTML = getOptionsFromArray(config.LLM_LOCAL_MODELS);
-    llmLocalModelSelect.value = config.LLM_LOCAL_MODEL;
+    llmLocalModelSelect.innerHTML = getOptionsFromArray([...config.LLM_LOCAL_MODELS, customModelKey]);
+
+    if (config.LLM_LOCAL_MODELS.includes(config.LLM_LOCAL_MODEL)) {
+        llmLocalModelSelect.value = config.LLM_LOCAL_MODEL;
+    } else {
+        llmLocalModelSelect.value = customModelKey;
+    }
 
     document.getElementById("llmUrl").value = config.LLM_URL;
     document.getElementById("llmApiKey").value = config.LLM_API_KEY;
@@ -109,6 +121,12 @@ function showConfig() {
     document.getElementById("postProcessingPrompt").value = config.POST_PROCESSING_PROMPT;
     // toggle the post-processing settings based on the checkbox
     togglePostProcessingSettings();
+
+    document.getElementById("minimumWordCountCheck").checked = config.MINIMUM_WORD_COUNT_CHECK;
+    document.getElementById("minimumWordCountLimit").value = config.MINIMUM_WORD_COUNT_LIMIT;
+    document.getElementById("translateToEnglish").checked = config.TRANSLATE_TO_ENGLISH;
+
+    toggleMinimumWordCountLimit();
 }
 
 
@@ -138,6 +156,17 @@ function toggleTranscriptionSettings() {
         showHideSettings(".transcriptionServerSettings", "hidden", "visible");
         formValidations('.transcription-local-form', true);
         formValidations('.transcription-server-form', false);
+
+        let transcriptionLocalModelSelect = document.getElementById("transcriptionLocalModel");
+
+        if (transcriptionLocalModelSelect.value === customModelKey) {
+            document.getElementById('transcriptionLocalModelCustom').value = config.TRANSCRIPTION_LOCAL_MODEL;
+            showHideSettings(".transcriptionLocalCustomSettings", "visible", "hidden");
+            formValidations('#transcriptionLocalModelCustom', true);
+        } else {
+            showHideSettings(".transcriptionLocalCustomSettings", "hidden", "visible");
+            formValidations('#transcriptionLocalModelCustom', false);
+        }
     } else {
         showHideSettings(".transcriptionLocalSettings", "hidden", "visible");
         showHideSettings(".transcriptionServerSettings", "visible", "hidden");
@@ -157,6 +186,17 @@ function toggleLLMSettings() {
         showHideSettings(".llmServerSettings", "hidden", "visible");
         formValidations('.llm-local-form', true);
         formValidations('.llm-server-form', false);
+
+        let llmLocalModelSelect = document.getElementById("llmLocalModel");
+
+        if (llmLocalModelSelect.value === customModelKey) {
+            document.getElementById('llmLocalModelCustom').value = config.LLM_LOCAL_MODEL;
+            showHideSettings(".llmLocalCustomSettings", "visible", "hidden");
+            formValidations('#llmLocalModelCustom', true);
+        } else {
+            showHideSettings(".llmLocalCustomSettings", "hidden", "visible");
+            formValidations('#llmLocalModelCustom', false);
+        }
     } else {
         showHideSettings(".llmLocalSettings", "hidden", "visible");
         showHideSettings(".llmServerSettings", "visible", "hidden");
@@ -196,18 +236,34 @@ function togglePostProcessingSettings() {
     }
 }
 
+function toggleMinimumWordCountLimit() {
+    const isChecked = document.getElementById("minimumWordCountCheck").checked;
+    const limitInput = document.getElementById("minimumWordCountLimit").parentElement;
+    limitInput.style.display = isChecked ? "block" : "none";
+}
+
 // Function: updateConfig - Update the configuration settings based on the form inputs
 // Save the configuration settings and close the tab
 function updateConfig() {
     // Transcription settings
     config.TRANSCRIPTION_LOCAL = document.getElementById("transcriptionLocal").checked;
     config.TRANSCRIPTION_LOCAL_MODEL = document.getElementById("transcriptionLocalModel").value;
+
+    if (config.TRANSCRIPTION_LOCAL_MODEL === customModelKey) {
+        config.TRANSCRIPTION_LOCAL_MODEL = document.getElementById("transcriptionLocalModelCustom").value;
+    }
+
     config.TRANSCRIPTION_URL = document.getElementById("transcriptionUrl").value;
     config.TRANSCRIPTION_API_KEY = document.getElementById("transcriptionApiKey").value;
 
     // LLM settings
     config.LLM_LOCAL = document.getElementById("llmLocal").checked;
     config.LLM_LOCAL_MODEL = document.getElementById("llmLocalModel").value;
+
+    if (config.LLM_LOCAL_MODEL === customModelKey) {
+        config.LLM_LOCAL_MODEL = document.getElementById("llmLocalModelCustom").value;
+    }
+
     config.LLM_URL = document.getElementById("llmUrl").value.replace(/\/$/, "");
     config.LLM_API_KEY = document.getElementById("llmApiKey").value;
 
@@ -227,6 +283,10 @@ function updateConfig() {
     // Post-processing settings
     config.POST_PROCESSING = document.getElementById("postProcessing").checked;
     config.POST_PROCESSING_PROMPT = document.getElementById("postProcessingPrompt").value;
+
+    config.MINIMUM_WORD_COUNT_CHECK = document.getElementById("minimumWordCountCheck").checked;
+    config.MINIMUM_WORD_COUNT_LIMIT = parseInt(document.getElementById("minimumWordCountLimit").value, 10);
+    config.TRANSLATE_TO_ENGLISH = document.getElementById("translateToEnglish").checked;
 
     // Save configuration
     saveConfig(config).then(function () {
@@ -356,9 +416,12 @@ document.addEventListener("DOMContentLoaded", async function (event) {
 
     // Event listeners for the checkboxes
     document.getElementById("transcriptionLocal").addEventListener("change", toggleTranscriptionSettings);
+    document.getElementById("transcriptionLocalModel").addEventListener("change", toggleTranscriptionSettings);
     document.getElementById("llmLocal").addEventListener("change", toggleLLMSettings);
+    document.getElementById("llmLocalModel").addEventListener("change", toggleLLMSettings);
     document.getElementById("preProcessing").addEventListener("change", togglePreProcessingSettings);
     document.getElementById("postProcessing").addEventListener("change", togglePostProcessingSettings);
+    document.getElementById("minimumWordCountCheck").addEventListener("change", toggleMinimumWordCountLimit);
 
     // listen to when the LLM server settings change and update the models
     document.getElementById("llmUrl").addEventListener("focusout", loadRemoteModelsOnSettingsChange);
