+++ src/popup.html
@@ -6,11 +6,46 @@
     <title>FreeScribe</title>
     <link rel="stylesheet" href="main.css"/>
     <link rel="stylesheet" href="fontawesome.min.css"/>
+    <style>
+.loading-overlay {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background-color: rgba(255, 255, 255, 0.9);
+    display: none;
+    justify-content: center;
+    align-items: center;
+    z-index: 1000;
+    flex-direction: column;
+}
+
+.loading-content {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 1rem;
+}
+
+.loading-text {
+    font-size: 1.2rem;
+    color: #333;
+}
+</style>
     <script src="jquery-3.7.1.min.js"></script>
     <script src="toastr.min.js"></script>
     <script src="bootstrap.bundle.min.js"></script>
 </head>
 <body>
+<div id="loadingOverlay" class="loading-overlay">
+    <div class="loading-content">
+        <div class="spinner-border text-primary" style="width: 3rem; height: 3rem;" role="status">
+            <span class="visually-hidden">Loading...</span>
+        </div>
+        <p id="loadingText" class="loading-text">Loading models...</p>
+    </div>
+</div>
 <div class="popup-container">
     <div class="container fade-in">
         <h3>Select Audio Input</h3>
@@ -50,6 +85,8 @@
             </button>
         </div>
 
+        <div id="errorMessage" class="alert alert-danger" role="alert" style="display: none;"></div>
+
         <div class="textarea-container mb-3">
           <textarea
                   id="userInput"
@@ -65,7 +102,7 @@
         </div>
 
         <button id="generateNotesButton" class="mb-3">
-            <i class="fas fa-file-alt"></i> Generate Notes
+            <i class="fas fa-file-alt"></i> Generate Note
         </button>
 
         <pre id="notes" class="mt-3"></pre>
@@ -134,6 +171,6 @@
     </div>
 </div>
 
-<script type="module" src="index.js"></script>
+<script type="module" src="popup.js"></script>
 </body>
 </html>
