+++ src/worker.js
@@ -1,16 +1,30 @@
+// Description: Web worker for running speech-to-text and language model tasks.
+// This file contains the code for the web worker that runs speech-to-text and language model tasks.
+// It listens for messages from the main thread and performs the required tasks.
+// It uses the transformers library to run the speech-to-text and language model tasks.
+// It sends messages back to the main thread with the results of the tasks.
+
 import {
-    pipeline, WhisperTextStreamer, InterruptableStoppingCriteria,
+    pipeline, WhisperTextStreamer, env
 } from "./transformers.min.js";
+
+// flag to prevent multiple transcriptions at once
+let isTranscribing = false;
+
+// Set wasmPaths to undefined to prevent remote loading of the WASM file.
+// The local WASM file is packaged into the build.
+// This resolves a loading error in the service worker.
+env.backends.onnx.wasm.wasmPaths = undefined;
+
+// Define message types
+const text2speech = "s2t";
+const llm = "llm";
 
 // Define model factories
 // Ensures only one model is created of each type
-
-const text2speech = "s2t";
-const llm = "llm";
-const stopping_criteria = new InterruptableStoppingCriteria();
-let isRecording = false;
-
-class TranslationPipeline {
+// provides progress callback to track model loading
+// and dispose of the model when it is no longer neededS
+class TranscriptionPipeline {
     static task = "automatic-speech-recognition";
     static model = "onnx-community/whisper-base";
     static instance = null;
@@ -29,12 +43,16 @@ class TranslationPipeline {
     }
 }
 
+// Function: loadSpeech2Text - Load the speech-to-text model.
+// Load the speech-to-text model and save it for future use.
+// Send messages to the main thread to track the progress of the model loading.
+// Send a message to the main thread when the model is loaded and ready.
 async function loadSpeech2Text(model) {
     self.postMessage({
         type: text2speech, status: "loading", message: "Loading model...",
     });
 
-    const p = TranslationPipeline;
+    const p = TranscriptionPipeline;
     if (p.model !== model) {
         // Invalidate model if different
         p.model = model;
@@ -55,39 +73,48 @@ async function loadSpeech2Text(model) {
     self.postMessage({type: text2speech, status: "ready"});
 }
 
-async function transcribe(audio) {
-    if (isRecording) {
+// Function: transcribe - Transcribe the audio to text.
+// Transcribe the audio to text using the speech-to-text model.
+// Send messages to the main thread to track the progress of the transcription.
+// Send a message to the main thread when the transcription is complete.
+// Prevent multiple transcriptions from running at once.
+// Chunk the audio and process it in chunks to improve performance.
+async function transcribe(data) {
+    if (isTranscribing) {
         return;
     }
-    isRecording = true;
 
+    const {audio, translate} = data;
+
+    isTranscribing = true;
+
+    // Tell the main thread we are starting
     self.postMessage({
         type: text2speech, status: "start",
     });
 
     // Load transcriber model
-    const transcriber = await TranslationPipeline.getInstance((data) => {
+    const transcriber = await TranscriptionPipeline.getInstance((data) => {
         data.type = text2speech;
         self.postMessage(data);
     });
 
+    // Define chunk and stride lengths in seconds
     const chunk_length_s = 30;
     const stride_length_s = 5;
-
     const time_precision = transcriber.processor.feature_extractor.config.chunk_length / transcriber.model.config.max_source_positions;
 
     // Storage for chunks to be processed. Initialise with an empty chunk.
     /** @type {{ text: string; offset: number, timestamp: [number, number | null] }[]} */
     const chunks = [];
 
-    // TODO: Storage for fully-processed and merged chunks
-    // let decoded_chunks = [];
-
+    // Variables to keep track of progress
     let chunk_count = 0;
     let start_time;
     let num_tokens = 0;
     let tps;
-    const streamer = new WhisperTextStreamer(transcriber.tokenizer, {
+
+    let transcriberOptions = {
         time_precision, on_chunk_start: (x) => {
             const offset = (chunk_length_s - stride_length_s) * chunk_count;
             chunks.push({
@@ -104,16 +131,6 @@ async function transcribe(audio) {
             }
             // Append text to the last chunk
             chunks.at(-1).text += x;
-
-            // self.postMessage({
-            //   type: text2speech,
-            //   status: "update",
-            //   data: {
-            //     text: "", // No need to send full text yet
-            //     chunks,
-            //     tps,
-            //   },
-            // });
         }, on_chunk_end: (x) => {
             const current = chunks.at(-1);
             current.timestamp[1] = x + current.offset;
@@ -123,7 +140,16 @@ async function transcribe(audio) {
             num_tokens = 0;
             ++chunk_count;
         },
-    });
+    }
+
+    if (translate) {
+        // Set the language model to use for transcription
+        transcriberOptions.language = "en";
+        transcriberOptions.task = "translate";
+    }
+
+    // Create a streamer to process the text in chunks
+    const streamer = new WhisperTextStreamer(transcriber.tokenizer, transcriberOptions);
 
     // Actually run transcription
     const output = await transcriber(audio, {
@@ -133,8 +159,6 @@ async function transcribe(audio) {
         // Sliding window
         chunk_length_s, stride_length_s,
 
-        // Language and task
-
         // Return timestamps
         return_timestamps: true, force_full_sequences: false,
 
@@ -142,21 +166,27 @@ async function transcribe(audio) {
         streamer, // after each generation step
     }).catch((error) => {
         console.error(error);
+        isTranscribing = false;
         self.postMessage({
             type: text2speech, status: "error", data: error,
         });
         return null;
     });
 
+    // Post the transcription back to the main thread
     self.postMessage({
         type: text2speech, status: "complete", data: {
             text: output.text, // No need to send full text yet
             chunks, tps,
         },
     });
-    isRecording = false;
+    isTranscribing = false;
 }
 
+// Define model factories
+// Ensures only one model is created of each type
+// provides progress callback to track model loading
+// and dispose of the model when it is no longer needed
 class LlmPipeline {
     static task = "text-generation";
     static model = "onnx-community/Phi-3.5-mini-instruct-onnx-web";
@@ -173,7 +203,12 @@ class LlmPipeline {
     }
 }
 
+// Function: loadLlm - Load the language model.
+// Load the language model and save it for future use.
+// Send messages to the main thread to track the progress of the model loading.
+// Send a message to the main thread when the model is loaded and ready.
 async function loadLlm(model) {
+    // Tell the main thread we are starting
     self.postMessage({
         type: llm, status: "loading", message: "Loading model...",
     });
@@ -197,9 +232,14 @@ async function loadLlm(model) {
         self.postMessage(x);
     });
 
+    // Tell the main thread we are ready
     self.postMessage({type: llm, status: "ready"});
 }
 
+// Function: generate - Generate text using the language model.
+// Generate text using the language model based on the input data.
+// Send messages to the main thread to track the progress of the generation.
+// Send a message to the main thread when the generation is complete.
 async function generate(data) {
     const {message, type, extra} = data;
     // Retrieve the text-generation pipeline.
@@ -208,10 +248,13 @@ async function generate(data) {
     // Tell the main thread we are starting
     self.postMessage({type: llm, status: "start"});
 
+    // Generate the prompt for the language model.
     const prompt = [{role: "user", content: message}];
 
+    // Generate the response using the language model.
     const result = await generator(prompt, {max_new_tokens: 128});
 
+    // Retrieve the generated text from the result.
     let outputText;
     try {
         const lastGenerated = result[0]?.generated_text?.at(-1);
@@ -223,6 +266,7 @@ async function generate(data) {
         outputText = "Failed to generate response";
     }
 
+    // Post the generated text back to the main thread.
     self.postMessage({
         type: llm, status: "complete", data: {
             type: type, text: outputText, extra: extra,
@@ -230,6 +274,7 @@ async function generate(data) {
     });
 }
 
+// Listen for messages from the main thread and perform the required tasks.
 self.addEventListener("message", async (event) => {
     const {type, data} = event.data;
 
@@ -244,16 +289,7 @@ self.addEventListener("message", async (event) => {
             loadLlm(data);
             break;
         case "generate":
-            stopping_criteria.reset();
             generate(data);
-            break;
-        case "interrupt":
-            stopping_criteria.interrupt();
-            break;
-
-        case "reset":
-            // past_key_values_cache = null;
-            stopping_criteria.reset();
             break;
     }
 });
